Add support ticket and bug reporting system
Implemented a support ticket and bug reporting system using newly added ButtonListener. Enhanced the DiscordBot to manage a list of tickets and included a ReadyListener to present the support and bug report options to users upon bot startup. The new system enables users to create tickets for support, bug reporting, and allows for tracking and closing of these tickets.